backend/pyproject.toml
[project]
name = "local-ai-knowledge-space-backend"
version = "0.1.0"
description = "Local-first AI knowledge workspace backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "aiofiles>=24.1.0",
  "pymupdf>=1.24.9",
  "pytesseract>=0.3.13",
  "pillow>=10.4.0",
  "chromadb>=0.5.5",
  "numpy>=2.0.0",
  "pydantic-ai>=1.0.0",
  "openai>=1.40.0",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=3.0.0"]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8"
]

[tool.ruff]
line-length = 100
target-version = "py312"

